Move PageRouter.js contents to App.js
